% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSet-methods.R, R/import-class.R
\name{coerce}
\alias{coerce}
\alias{coerce,BiocSet,list-method}
\alias{coerce,BiocSet,GeneSetCollection-method}
\title{as("BiocSet", "list")}
\description{
as("BiocSet", "list")

as("BiocSet", "GeneSetCollection")
}
