% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ElementSet.R
\docType{class}
\name{ElementSet-class}
\alias{ElementSet-class}
\alias{.ElementSet}
\alias{ElementSet}
\alias{es_activate}
\alias{es_activate.ElementSet}
\alias{update_es_element}
\alias{update_es_set}
\alias{update_es_elementset}
\alias{es_element}
\alias{es_element,ElementSet-method}
\alias{es_set}
\alias{es_set,ElementSet-method}
\alias{es_elementset}
\alias{es_elementset,ElementSet-method}
\alias{es_element<-}
\alias{es_set<-}
\alias{es_elementset<-}
\alias{filter.ElementSet}
\alias{select.ElementSet}
\alias{mutate.ElementSet}
\alias{map_element}
\alias{map_element.ElementSet}
\alias{map_set}
\alias{map_set.ElementSet}
\alias{group_by.ElementSet}
\alias{summarise.ElementSet}
\alias{arrange.ElementSet}
\alias{tbl_vars.ElementSet}
\title{An element set representation as a tripple tibble}
\usage{
ElementSet(..., active = c("elementset", "element", "set"))

es_activate(.data, what)

\method{es_activate}{ElementSet}(.data, what)

update_es_element(es, value)

update_es_set(es, value)

update_es_elementset(es, value)

es_element(x)

\S4method{es_element}{ElementSet}(x)

es_set(x)

\S4method{es_set}{ElementSet}(x)

es_elementset(x)

\S4method{es_elementset}{ElementSet}(x)

es_element(es) <- value

es_set(es) <- value

es_elementset(es) <- value

\method{filter}{ElementSet}(.data, ...)

\method{select}{ElementSet}(.data, ...)

\method{mutate}{ElementSet}(.data, ...)

map_element(.data, from, to)

\method{map_element}{ElementSet}(.data, from, to)

map_set(.data, from, to)

\method{map_set}{ElementSet}(.data, from, to)

\method{group_by}{ElementSet}(.data, ..., add = FALSE)

\method{summarise}{ElementSet}(.data, ...)

\method{arrange}{ElementSet}(.data, ...)

\method{tbl_vars}{ElementSet}(x)
}
\arguments{
\item{...}{For `ElementSet()`, named character() vectors of element
sets. Each character vector is an element set. The name of the
character vector is the name of the element set.}

\item{active}{A character to indicate which tibble is active.}

\item{.data}{The 'ElementSet' tibble.}

\item{what}{Which of the three tibbles to activate}

\item{es}{The active tibble}

\item{value}{What it is being updated to}

\item{x}{The active tibble}

\item{from}{a vector of the values to be replaced}

\item{to}{a vector of the replacement values}

\item{add}{by default, `group_by()` will override existing groups. To add to
existing groups, add should be TRUE.}
}
\value{
For `ElementSet()`, an S4 'ElementSet' object in a tripple
    tibble representation.
}
\description{
An element set representation as a tripple tibble
}
\section{Slots}{

\describe{
\item{\code{element}}{The element tibble from `tbl_elementset`}

\item{\code{set}}{The set tibble from `tbl_elementset`}

\item{\code{elementset}}{The elementset tibble created from user input}

\item{\code{active}}{The tibble that is active}
}}

\examples{
ElementSet(set1 = letters, set2 = LETTERS)
es <- ElementSet(set1 = letters, set2 = LETTERS)
es_activate(es, element)
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(element) \%>\% filter(element == "a")
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% select(element)
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(set) \%>\% mutate(pval = rnorm(1:2))
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% map_element(letters, LETTERS)
es <- ElementSet(a = letters, B = LETTERS)
es \%>\% map_set("set1", "foo")
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% group_by(element, set)
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(set) \%>\% summarise(n = n())
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(element) \%>\% arrange(desc(element))
es <- ElementSet(set1 = letters, set2 = LETTERS)
es \%>\% mutate(pval = rnorm(1:52)) \%>\% tbl_vars()
}
