% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblgene-class.R
\name{tbl_gene}
\alias{tbl_gene}
\title{Gene representation as an S3 class tibble}
\usage{
tbl_gene(tbl_geneset)
}
\arguments{
\item{tbl_geneset}{An S3 'geneset' tibble.}
}
\value{
An S3 'gene' object in a tibble representation.
}
\description{
Gene representation as an S3 class tibble
}
\examples{
tbl <- tbl_geneset(set1 = letters, set2 = LETTERS)
tbl_gene(tbl)
}
