% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-GeneSetCollection.R
\name{genesetcollection}
\alias{genesetcollection}
\alias{GeneSetCollection_from_BiocSet}
\alias{BiocSet_from_GeneSetCollection}
\title{GeneSetCollection}
\usage{
GeneSetCollection_from_BiocSet(biocset)

BiocSet_from_GeneSetCollection(gsc)
}
\arguments{
\item{biocset}{The \code{BiocSet} object that will become a
\code{GeneSetCollection} object.}

\item{gsc}{The \code{GeneSetCollection} that will become a \code{BiocSet}
object.}
}
\value{
For `GeneSetCollection_from_BiocSet()`, a GeneSetCollection.

For `BiocSet_from_GeneSetCollection()`, a BiocSet object.
}
\description{
The following functions deal with converting a \code{BiocSet}
    object into a \code{GeneSetCollection} object, or vice versa.
}
\examples{
biocset <- BiocSet(set1 = letters, set2 = LETTERS)
gsc <- GeneSetCollection_from_BiocSet(biocset)
gsc

BiocSet_from_GeneSetCollection(gsc)
}
