% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetS3tbl-class.R
\name{GeneSetS3tbl}
\alias{GeneSetS3tbl}
\alias{names.GeneSetS3tbl}
\alias{lengths.GeneSetS3tbl}
\alias{print.GeneSetS3tbl}
\title{Gene set representation as in S3 class tibble}
\usage{
GeneSetS3tbl(...)

\method{names}{GeneSetS3tbl}(x)

lengths.GeneSetS3tbl(x, use.names = TRUE)

\method{print}{GeneSetS3tbl}(x, ...)
}
\arguments{
\item{...}{For `GeneSetS3tbl()`, named character() vectors of gene sets. Each
character vector is a gene set. The name of the character vector is the
name of the gene set.}

\item{x}{An S3 'GeneSetS3tbl' object}

\item{use.names}{Logical indicating if the result should inherit the names
from x.}
}
\value{
For `GeneSetS3tbl()`, an S3 'GeneSetS3' object in a tibble
   representation.
}
\description{
Gene set representation as in S3 class tibble
}
\examples{
GeneSetS3tbl(set1 = letters, set2 = LETTERS)
}
