% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-class.R
\docType{class}
\name{import.gmt}
\alias{import.gmt}
\alias{GMTFile-class}
\alias{.GMTFile}
\alias{GMTFile}
\alias{import,GMTFile,ANY,ANY-method}
\alias{export.tbl_geneset}
\alias{export,tbl_geneset,GMTFile,ANY-method}
\title{Importing and formating of gene sets as an S3 class tibble}
\usage{
import.gmt(path)

GMTFile(resource, ...)

\S4method{import}{GMTFile,ANY,ANY}(con, format, text, ...)

export.tbl_geneset(tbl, path = tempfile(fileext = ".gmt"))

\S4method{export}{tbl_geneset,GMTFile,ANY}(object, con, format, ...)
}
\arguments{
\item{path}{For `import.gmt()`, a file name or URL containing gene sets.}

\item{tbl}{For `export.tbl_geneset()`, a tbl_geneset that
should be exported to a gmt file.}
}
\description{
Importing and formating of gene sets as an S3 class tibble
}
\examples{
gmtFile <- system.file(package = "GeneSet", "extdata",
    "hallmark.gene.symbol.gmt")
import.gmt(gmtFile)
}
