% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_single.R
\name{intersect_single}
\alias{intersect_single}
\title{Intersect on a single \code{BiocSet} object}
\usage{
intersect_single(x, ...)
}
\arguments{
\item{x}{A \code{BiocSet} object.}

\item{...}{Additional arguments passed to function.}
}
\value{
A \code{BiocSet} object with a single set  'intersect' and 
    interesected elements from x.
}
\description{
This function performs an intersection within a single 
    \code{BiocSet} object.
}
\examples{
\dontrun{
es1 <- BiocSet(set1 = letters[c(1:10)], set2 = letters[c(4:20)])
intersect_single(es1)
}
}
