% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSet.R
\name{filter.BiocSet}
\alias{filter.BiocSet}
\alias{select.BiocSet}
\alias{mutate.BiocSet}
\alias{summarise.BiocSet}
\alias{arrange.BiocSet}
\alias{tbl_nongroup_vars.BiocSet}
\alias{group_by.BiocSet}
\alias{left_join_element}
\alias{left_join_set}
\title{Major dplyr functions adapted into BiocSet}
\usage{
\method{filter}{BiocSet}(.data, ...)

\method{select}{BiocSet}(.data, ...)

\method{mutate}{BiocSet}(.data, ...)

\method{summarise}{BiocSet}(.data, ...)

\method{arrange}{BiocSet}(.data, ...)

tbl_nongroup_vars.BiocSet(x)

\method{group_by}{BiocSet}(.data, ..., add = FALSE)

left_join_element(.data, ...)

left_join_set(.data, ...)
}
\arguments{
\item{.data}{The 'BiocSet' tibble}

\item{...}{Additional arguments passed to function.}

\item{x}{A 'BiocSet' object.}

\item{add}{logical, whether to add to the existing groups.}
}
\value{
A 'BiocSet' object
}
\description{
Major dplyr functions adapted into BiocSet
}
\examples{
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(element) \%>\% filter(element == "a")
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% select(element)
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(set) \%>\% mutate(pval = rnorm(1:2))
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(set) \%>\% summarise(n = n())
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(element) \%>\% arrange(desc(element))
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% mutate(pval = rnorm(1:52)) \%>\% tbl_nongroup_vars()
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% group_by(element, set)
tbl <- tibble(x = 1:5, y = letters[1:5])
es <- BiocSet(set1 = letters[c(1,3,5)], set2 = letters[c(2,4)])
es \%>\% left_join_element(tbl, by = c(element = "y"))
tbl <- tibble(x = 10:11, y = c("set1", "set2"))
es <- BiocSet(set1 = letters[c(1,3,5)], set2 = letters[c(2,4)])
es \%>\% left_join_set(tbl, by = c(set = "y"))
}
