% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{go_sets}
\alias{go_sets}
\alias{es_map}
\alias{kegg_sets}
\title{Functions for mapping elements and sets to different ids}
\usage{
go_sets(org, from)

es_map(es, org, from, to)

kegg_sets(species, pathways)
}
\arguments{
\item{org}{The AnnotationDbi object to identify keys/mappings from}

\item{from}{A character to indicate which identifier to map from}

\item{es}{The BiocSet object to map the elements on}

\item{to}{A character to indicate which identifier to map to}

\item{species}{Which species the pathways are from}

\item{pathways}{A character vector containing the pathways to map from}
}
\value{
A BiocSet object with Entrez IDs reported as elements (default from
       KEGGREST) and KEGG pathways as sets.
}
\description{
Functions for mapping elements and sets to different ids
}
\examples{
library(org.Hs.eg.db)
go_sets(org.Hs.eg.db, "ENSEMBL")
library(org.Hs.eg.db)
es <- BiocSet(set1 = c("PRKACA", "TGFA", "MAP2K1"), set2 = c("CREB3", "FOS"))
es_map(es, org.Hs.eg.db, "SYMBOL", "ENTREZID") 
pathways <- c("hsa05310", "hsa04110", "hsa05224", "hsa04970")
kegg_sets("hsa", pathways)
}
