% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetS3-class.R
\name{print.GeneSetS3}
\alias{print.GeneSetS3}
\title{Print method for the S3 class gene set representation}
\usage{
\method{print}{GeneSetS3}(x)
}
\arguments{
\item{...}{An S3 'GeneSetS3' object}
}
\value{
A list of descriptive characteristics of the object
}
\description{
Print method for the S3 class gene set representation
}
\examples{
GeneSetS3(set1 = letters, set2= LETTERS)
}
