% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_set.R
\name{tbl_set}
\alias{tbl_set}
\title{Set representation as an S3 class tibble}
\usage{
tbl_set(tbl_geneset)
}
\arguments{
\item{tbl_geneset}{An S3 'geneset' tibble.}
}
\value{
An S3 'set' object in a tibble representation.
}
\description{
Set representation as an S3 class tibble
}
\examples{
tbl <- tbl_geneset(set1 = letters, set2 = LETTERS)
tbl_set(tbl)
}
