% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblelement-class.R
\name{tbl_element}
\alias{tbl_element}
\title{Element representation as an S3 class tibble}
\usage{
tbl_element(tbl_elementset)
}
\arguments{
\item{tbl_elementset}{An S3 'elementset' tibble.}
}
\value{
An S3 'element' object in a tibble representation.
}
\description{
Element representation as an S3 class tibble
}
\examples{
tbl <- tbl_elementset(set1 = letters, set2 = LETTERS)
tbl_element(tbl)
}
