% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblset-class.R
\name{tblset}
\alias{tblset}
\alias{tbl_set}
\title{Set representation as an S3 class tibble}
\usage{
tbl_set(tbl_elementset)
}
\arguments{
\item{tbl_elementset}{An S3 \code{elementset} tibble.}
}
\value{
An S3 \code{set} object in a tibble representation.
}
\description{
Set representation as an S3 class tibble
}
\examples{
tbl <- tbl_elementset(set1 = letters, set2 = LETTERS)
tbl_set(tbl)
}
