% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_single.R
\name{union_single}
\alias{union_single}
\title{Union on a single \code{BiocSet} object}
\usage{
union_single(x, ...)
}
\arguments{
\item{x}{A \code{BiocSet} object.}

\item{...}{Additional arguments passed to function.}
}
\value{
For \code{union_single}, a \code{BiocSet} object with a single set 
    \code{union} and unioned elements from x.
}
\description{
This function performs a union within a single \code{BiocSet}
    object.
}
\examples{
es3 <- BiocSet(set1 = letters[c(1:10)], set2 = letters[c(4:20)])
union_single(es3)
}
