% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblgeneset-class.R
\name{tbl_geneset}
\alias{tbl_geneset}
\alias{format.trunc_mat_tbl_geneset}
\title{Gene set representation as an S3 class tibble}
\usage{
tbl_geneset(...)

\method{format}{trunc_mat_tbl_geneset}(x, ...)
}
\arguments{
\item{...}{For `tbl_geneset()`, named character() vectors of gene
sets. Each character vector is a gene set. The name of the
character vector is the name of the gene set.}

\item{x}{An object of class trunc_mat_tbl_geneset, used during
printing of tbl_geneset}
}
\value{
For `tbl_geneset()`, an S3 'geneset' object in a tibble
    representation.
}
\description{
Gene set representation as an S3 class tibble
}
\examples{
tbl_geneset(set1 = letters, set2 = LETTERS)
}
