% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSet.R
\name{GeneSet}
\alias{GeneSet}
\alias{gs_activate.GeneSet}
\alias{filter.GeneSet}
\alias{select.GeneSet}
\alias{mutate.GeneSet}
\alias{group_by.GeneSet}
\alias{ungroup.GeneSet}
\alias{summarise.GeneSet}
\alias{arrange.GeneSet}
\title{A gene set representation as a tripple tibble}
\usage{
GeneSet(..., active = c("geneset", "gene", "set"))

\method{gs_activate}{GeneSet}(.data, what)

\method{filter}{GeneSet}(.data, ...)

\method{select}{GeneSet}(.data, ...)

\method{mutate}{GeneSet}(.data, ...)

\method{group_by}{GeneSet}(.data, ..., add = FALSE)

\method{ungroup}{GeneSet}(.data, ...)

\method{summarise}{GeneSet}(.data, ...)

\method{arrange}{GeneSet}(.data, ...)
}
\arguments{
\item{...}{For `GeneSet()`, named character() vectors of gene
sets. Each character vector is a gene set. The name of the
character vector is the name of the gene set.}

\item{active}{A character to indicate which tibble is active.}

\item{.data}{The 'GeneSet' tibble.}

\item{what}{Which of the three tibbles to activate}
}
\value{
For `GeneSet()`, a S4 'GeneSet' object in a tripple
    tibble representation.
}
\description{
A gene set representation as a tripple tibble
}
\examples{
GeneSet(set1 = letters, set2 = LETTERS)
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs_activate(gs, gene)
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% gs_activate(gene) \%>\% filter(gene == "a")
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% select(gene)
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% gs_activate(set) \%>\% mutate(pval = rnorm(1:2))
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% group_by(gene, set)
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% group_by(set) \%>\% summarise(n = n()) \%>\% ungroup()
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% gs_activate(set) \%>\% summarise(n = n())
gs <- GeneSet(set1 = letters, set2 = LETTERS)
gs \%>\% gs_activate(gene) \%>\% arrange(desc(gene))
}
