% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetS3-class.R
\name{GeneSetS3}
\alias{GeneSetS3}
\alias{print.GeneSetS3}
\alias{[.GeneSetS3}
\alias{GeneSetS3_tbl}
\alias{names.GeneSetS3_tbl}
\alias{lengths.GeneSetS3_tbl}
\alias{print.GeneSetS3_tbl}
\title{Gene set representation as an S3 class}
\usage{
GeneSetS3(...)

\method{print}{GeneSetS3}(x, ...)

\method{[}{GeneSetS3}(x, i, j, ..., drop = TRUE)

GeneSetS3_tbl(...)

\method{names}{GeneSetS3_tbl}(x)

lengths.GeneSetS3_tbl(x, use.names = TRUE)

\method{print}{GeneSetS3_tbl}(x, ...)
}
\arguments{
\item{...}{For `GeneSetS3()`, named character() vectors of gene sets. Each
character vector is a gene set. The name of the character vector is the
name of the gene set.}

\item{x}{An S3 'GeneSetS3' object}

\item{i}{A numeric or character vector, indicy specifying elements to
extract or replace}

\item{j}{A numeric or character vector, indicy specifying elements to
extract or replace}

\item{drop}{logical, if TRUE the result is coerced to lowest possible
dimension. Only works for extraction, not replacement}

\item{use.names}{Logical indicating if the result should inherit the names
from x.}
}
\value{
For `GeneSetS3()`, an S3 'GeneSetS3' object.

For `GeneSetS3_tbl()`, an S3 'GeneSetS3' object in a tibble
   representation.
}
\description{
Gene set representation as an S3 class
}
\examples{
GeneSetS3(set1 = c("A", "B", "C"), set2 = c("D", "E"))
}
