% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetS3-class.R
\name{GeneSetS3}
\alias{GeneSetS3}
\title{Gene set representation as an S3 class}
\usage{
GeneSetS3(...)
}
\arguments{
\item{...}{named character() vectors of gene sets. Each character vector is
a gene set. The name of the character vector is the name of the gene set.}
}
\value{
An S3 'GeneSetS3' object.
}
\description{
Gene set representation as an S3 class
}
\examples{
GeneSetS3(set1 = c("A", "B", "C"), set2 = c("D", "E"))
}
