% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetS4-class.R
\docType{class}
\name{GeneSetDf-class}
\alias{GeneSetDf-class}
\alias{.GeneSetDf}
\title{Gene set representation as an S4 class}
\description{
Gene set representation as an S4 class
}
\section{Slots}{

\describe{
\item{\code{gene}}{charcter() vector of genes}

\item{\code{set}}{factor() of gene sets}
}}

