% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-class.R
\docType{class}
\name{import.gmt}
\alias{import.gmt}
\alias{GMTFile-class}
\alias{.GMTFile}
\alias{GMTFile}
\alias{import,GMTFile,ANY,ANY-method}
\alias{export.tbl_geneset}
\alias{export,tbl_geneset,GMTFile,ANY-method}
\title{Importing and formating of gene sets as an S3 class tibble}
\usage{
import.gmt(path)

GMTFile(resource, ...)

\S4method{import}{GMTFile,ANY,ANY}(con, format, text, ...)

export.tbl_geneset(tbl, path = tempfile(fileext = ".gmt"))

\S4method{export}{tbl_geneset,GMTFile,ANY}(object, con, format, ...)
}
\arguments{
\item{path}{For `import.gmt()`, a file name or URL containing gene sets.}

\item{resource}{For `GMTFile()`, the .gmt file that will be imported in.}

\item{...}{Parameters to pass to the format-specific method}

\item{con}{For `import()`, the file name or URL the gene set is
loaded from. For `export()`, the file name or URL the gene set
is to be saved to.}

\item{format}{For `import()`, the format of the output.}

\item{text}{For `import()`, if con is missing this is a character
vector directly providing the gene set that should be imported.}

\item{tbl}{For `export.tbl_geneset()`, a tbl_geneset that
should be exported to a gmt file.}

\item{object}{For `export()`, the object to be exported.}
}
\description{
Importing and formating of gene sets as an S3 class tibble
}
\examples{
gmtFile <- system.file(package = "GeneSet", "extdata",
    "hallmark.gene.symbol.gmt")
import.gmt(gmtFile)
}
