% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSet.R
\name{filter.BiocSet}
\alias{filter.BiocSet}
\alias{select.BiocSet}
\alias{mutate.BiocSet}
\alias{summarise.BiocSet}
\alias{arrange.BiocSet}
\alias{tbl_nongroup_vars.BiocSet}
\alias{group_by.BiocSet}
\alias{left_join.BiocSet}
\title{Major dplyr functions adapted into BiocSet}
\usage{
\method{filter}{BiocSet}(.data, ...)

\method{select}{BiocSet}(.data, ...)

\method{mutate}{BiocSet}(.data, ...)

\method{summarise}{BiocSet}(.data, ...)

\method{arrange}{BiocSet}(.data, ...)

tbl_nongroup_vars.BiocSet(x)

\method{group_by}{BiocSet}(.data, ..., add = FALSE)

\method{left_join}{BiocSet}(.data, ...)
}
\arguments{
\item{.data}{The 'BiocSet' tibble}

\item{...}{Additional arguments passed to function.}

\item{x}{A 'BiocSet' object.}

\item{add}{logical, whether to add to the existing groups.}
}
\value{
A 'BiocSet' object
}
\description{
Major dplyr functions adapted into BiocSet
}
\examples{
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(element) \%>\% filter(element == "a")
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% select(element)
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(set) \%>\% mutate(pval = rnorm(1:2))
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(set) \%>\% summarise(n = n())
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% es_activate(element) \%>\% arrange(desc(element))
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% mutate(pval = rnorm(1:52)) \%>\% tbl_nongroup_vars()
es <- BiocSet(set1 = letters, set2 = LETTERS)
es \%>\% group_by(element, set)
es <- BiocSet(set1 = letters[1:5], set2 = LETTERS[1:5])
tbl <- tibble(x = 1:10, y = c(letters[1:5], LETTERS[1:5]))
es \%>\% left_join(tbl, by = c(element = "y"))
}
