% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{go_sets}
\alias{go_sets}
\alias{es_map_unique}
\alias{es_map_multiple}
\alias{kegg_sets}
\alias{map_add_element}
\alias{map_add_set}
\title{Functions for mapping elements and sets to different ids}
\usage{
go_sets(org, from, go = c("GO", "GOID"), evidence = NULL,
  ontology = NULL)

es_map_unique(es, org, from, to)

es_map_multiple(es, org, from, to, multi = c("list", "filter", "asNA",
  "CharacterList", "FUN"))

kegg_sets(species)

map_add_element(es, org, from, add)

map_add_set(es, org, from, add)
}
\arguments{
\item{org}{The AnnotationDbi object to identify keys/mappings from}

\item{from}{A character to indicate which identifier to map from}

\item{go}{Character, the column name for the GO ids. Default is "GO".}

\item{evidence}{Character, the evidence codes for GO associations with 
a gene of interest. Default is all possible evidence codes.}

\item{ontology}{Character, which Gene Ontology. Default is BP, CC, and MF}

\item{es}{The BiocSet objec to map the elements on}

\item{to}{A character to indicate which identifier to map to}

\item{multi}{How should multiple values be returned? Options include:
list: This will just return a list object to the end user
filter: This will remove all elements that contain multiple matches and will
therefore return a shorter vector than what came in whenever some of the keys
match more than one value
asNA: This will return an NA value whenever there are multiple matches
CharacterList: This just returns a SimpleCharacterList object
FUN: You can also supply a function to the 'multiVals' argument for custom
behaviours.}

\item{species}{Which species the pathways are from}

\item{add}{The id to add to the `BiocSet` object}
}
\value{
A BiocSet object

For kegg_sets(), a BiocSet object with Entrez IDs reported as 
    elements (default from KEGGREST) and KEGG pathways as sets.
}
\description{
Functions for mapping elements and sets to different ids
}
\examples{
library(org.Hs.eg.db)
go_sets(org.Hs.eg.db, "ENSEMBL")
es <- BiocSet(set1 = c("PRKACA", "TGFA", "MAP2K1"), set2 = c("CREB3", "FOS"))
es_map_unique(es, org.Hs.eg.db, "SYMBOL", "ENTREZID")
es <- BiocSet(set1 = c("PRKACA", "TGFA", "MAP2K1"), set2 = c("CREB3", "FOS"))
es_map_mulitple(es, org.Hs.eg.db, "SYMBOL", "ENTREZID")
kegg_sets("hsa")
es <- BiocSet(set1 = c("PRKACA", "TGFA", "MAP2K1"), set2 = c("FOS", "BRCA1"))
map <- map_add_element(es, org.Hs.eg.db, "SYMBOL", "ENTREZID")
es \%>\% mutate_element(entrez = map)
library(GO.db)
go <- go_sets(org.Hs.eg.db, "ENSEMBL")
map <- map_add_set(go, GO.db, "GOID", "DEFINITION")
go \%>\% mutate_set(definition = map)
}
