% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{go_sets}
\alias{go_sets}
\alias{es_map}
\alias{kegg_sets}
\alias{map_add}
\title{Functions for mapping elements and sets to different ids}
\usage{
go_sets(org, from)

es_map(es, org, from, to)

kegg_sets(species)

map_add(es, org, from, add)
}
\arguments{
\item{org}{The AnnotationDbi object to identify keys/mappings from}

\item{from}{A character to indicate which identifier to map from}

\item{es}{The BiocSet object to map the elements on}

\item{to}{A character to indicate which identifier to map to}

\item{species}{Which species the pathways are from}

\item{add}{The id to add to the `BiocSet` object}
}
\value{
A BiocSet object

For kegg_sets(), a BiocSet object with Entrez IDs reported as 
    elements (default from KEGGREST) and KEGG pathways as sets.
}
\description{
Functions for mapping elements and sets to different ids
}
\examples{
library(org.Hs.eg.db)
go_sets(org.Hs.eg.db, "ENSEMBL")
es <- BiocSet(set1 = c("PRKACA", "TGFA", "MAP2K1"), set2 = c("CREB3", "FOS"))
es_map(es, org.Hs.eg.db, "SYMBOL", "ENTREZID") 
\dontrun{
kegg_sets("hsa")i
}
library(org.Hs.eg.db)
es <- BiocSet(set1 = c("PRKACA", "TGFA", "MAP2K1"), set2 = c("FOS", "BRCA1"))
map <- map_add(es, org.Hs.eg.db, "SYMBOL", "ENTREZID")
es \%>\% mutate_element(entrez = map)
}
