% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ops.R
\name{union.BiocSet}
\alias{union.BiocSet}
\alias{intersect.BiocSet}
\title{Set operations for BiocSets}
\usage{
\method{union}{BiocSet}(x, y = NULL, ...)

\method{intersect}{BiocSet}(x, y = NULL, ...)
}
\arguments{
\item{x}{The first BiocSet to perform operations on}

\item{y}{The second BiocSet to perform operations on}

\item{...}{other arguments passed on to methods}
}
\value{
For union, a tibble with the results of the union of x and y

For intersect, a tibble with the intersect of x and y
}
\description{
Set operations for BiocSets
}
\examples{
es1 <- BiocSet(set1 = letters[c(1:3)], set2 = LETTERS[c(1:3)])
es2 <- BiocSet(set1 = letters[c(2:4)], set2 = LETTERS[c(2:4)])
union(es1, es2)
es3 <- BiocSet(set1 = letters[c(1:10)], set2 = letters[c(4:20)])
union(es3)
es1 <- BiocSet(set1 = letters[c(1:4)], set2 = LETTERS[c(1:4)])
es2 <- BiocSet(set1 = letters[c(3:8)], set2 = LETTERS[c(3:8)])
intersect(es1, es2)
es3 <- BiocSet(set1 = letters[c(1:10)], set2 = letters[c(4:20)])
intersect(es3)
}
