% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblelementset-class.R
\name{tbl_elementset}
\alias{tbl_elementset}
\alias{format.trunc_mat_tbl_elementset}
\title{Element set representation as an S3 class tibble}
\usage{
tbl_elementset(...)

\method{format}{trunc_mat_tbl_elementset}(x, ...)
}
\arguments{
\item{...}{For `tbl_elementset()`, named character() vectors of gene
sets. Each character vector is a element set. The name of the
character vector is the name of the element set.}

\item{x}{An object of class trunc_mat_tbl_elementset, used during
printing of tbl_elementset}
}
\value{
For `tbl_elementset()`, an S3 'elementset' object in a tibble
    representation.
}
\description{
Element set representation as an S3 class tibble
}
\examples{
tbl_elementset(set1 = letters, set2 = LETTERS)
}
