% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_ref.R
\name{url_ref_element}
\alias{url_ref_element}
\alias{url_ref_set}
\alias{url_ref}
\title{Functions to access reference urls for different identifiers}
\usage{
url_ref_element(es)

url_ref_set(es)

url_ref(es)
}
\arguments{
\item{es}{A BiocSet object in which the reference urls should be added to.}
}
\value{
For `url_ref_element()`, a BiocSet object with the url column added 
    to the element tibble.

For `url_ref_set()`, a BiocSet object with the url column added to 
    the set tibble.

For `url_ref()`, a BiocSet object with the url column added to both 
    the element and set tibbles.
}
\description{
Functions to access reference urls for different identifiers
}
\examples{
es <- BiocSet("GO:0000002" = c("TP53", "TNF"), "GO:0000003" = c("IL6"))
url_ref_element(es) 

url_ref_set(es)

url_ref(es)
}
