% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_ref.R
\name{url_ref}
\alias{url_ref}
\title{Functions to access help for the different identifiers}
\usage{
url_ref(search, ...)
}
\arguments{
\item{search}{character, the id to look up. For KEGG, "map" will have to 
replace the letters before the numbers, e.g. "hsa05310" should
become "map05310".}

\item{...}{additional arguments to pass to 'browseURL'.}
}
\description{
Functions to access help for the different identifiers
}
\examples{
url_ref("GO:0000002")
}
