% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_ref.R
\name{url_ref_element}
\alias{url_ref_element}
\alias{url_ref_set}
\alias{url_ref}
\title{Functions to access help for the different identifiers}
\usage{
url_ref_element(es)

url_ref_set(es)

url_ref(es)
}
\arguments{
\item{es}{A BiocSet object in which the reference urls should be added to.}
}
\value{
A BiocSet object with the url column added to element or set.
}
\description{
Functions to access help for the different identifiers
}
\examples{
es <- BiocSet(set1 = c("TP53", "TNF", "EGFR"), set2 = c("IL6", "VEGFA"))
url_ref_element(es)
es <- BiocSet("GO:0000002" = c("TP53", "TNF"), "GO:0000003" = c("IL6")) 
url_ref_set(es)
es <- BiocSet("GO:0000002" = c("TP53", "TNF"), "GO:0000003" = c("IL6"))
url_ref(es)
}
